requirements.txt
alembic==1.3.2
aniso8601==8.0.0
attrs==19.3.0
autoenv==1.0.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
jwcrypto==0.6.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
pyparsing==2.4.5
pytest==5.3.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.7
Werkzeug==0.16.0
